Bug 1847319 - Block QCOM_tiled_rendering on all driver version on Adreno 308 GPUs. r=gfx-reviewers,lsalzman
In bug 1843749 we received reports of QCOM_tiled_rendering causing
glitches on Adreno 308 GPUs. As using this extension is a performance
win, especially on less powerful GPUs such as the Adreno 308, we
decided to block it only on the driver versions which we knew were
broken: V331 and V415.

However, we have now received a report that it is additionally broken
on version V0502. As it now seems likely the bug affects all driver
versions, or at least more than originally hoped, this patch blocks
the extension on the Adreno 308 regardless of driver version.

Differential Revision: https://phabricator.services.mozilla.com/D185530

UltraBlame original commit: 4297ef6efd09fb8021fdc7b4623f11e360e9f775