Bug 1846983 - Fix secondary text color in rich suggestions and make some other related fixes. r=dao
UX calls for this text to be a "secondary" color called
`--light-theme-text-secondary` and defined as `#5B5B66` (in light mode), which
is used in Firefox in a bunch of places [1], and in particular it's called
`--color-gray-70` in the design tokens CSS [2].

This patch uses a 73% color mix of the current color and transparent, which is
as close as I could get to `#5B5B66` in light mode. Using a mix lets this text
look nice on all themes.

Other fixes:

* Remove `--urlbar-popup-action-color` on root in urlbarView.css since it's
  always overridden and that color is never used AFAICT
* Remove an unused `--urlbar-popup-action-color: #30e60b;` rule on
  `:root[lwt-toolbar-field-focus-brighttext]`
* Fix selection and hover colors for action and URLs in rich suggestions

[1] https://searchfox.org/mozilla-central/search?q=5B5B66&case=false&regexp=false&path=
[2] https://searchfox.org/mozilla-central/rev/b4dadcc4c45a2e381a4866d28ebf3a901decccc1/toolkit/themes/shared/design-tokens-shared.css#22

Differential Revision: https://phabricator.services.mozilla.com/D185767

UltraBlame original commit: a5e815e81cc15bfa1139b51693f3979c28ffb4b8