Bug 1845257 - Prevent adding extra side-effect annotations r=iain
When an object is handled by Scalar Replacement, it might be annotated
as having side-effect, such side-effect are usually encoded as part of
the resume point to restore what the actual value was.

Unfortunately, this flag is stored on the object which has side-effect
and if the side-effect are localized to a small scope, not all Resume
Points might have a need side-effects to properly recover the content
of the object.

If one object were to be annotated, but the resume point happen to not
have any side-effect registered, then this would cause a problem where
the value would be replcade by an optimized-out magic-value instead of
being decoded.

This change makes it such that if an allocation has a flag which
explicitly request for a side-effect, but the ResumePoint does not
have any registered, then the allocation flag is ignored.

Differential Revision: https://phabricator.services.mozilla.com/D185579

UltraBlame original commit: a061e8b68e5578db9bc053b6ffc4f75550cc172f