Bug 1848304 - A bad TRR URL shows unexpected error in TRR mode3 error page r=necko-reviewers,fluent-reviewers,flod,kershaw
This patch adds a new TRR skip reason used when the TRR request is made
with an invalid URL. Normally this is a URL that doesn't parse, but if
the URL is a non-HTTPS URL, then the TRR service will just use an empty
string as the URL, leading to the same failure to parse it.

This skip reason will be reported when the page load is triggered.
We should report the cause to the user instead of just saying it failed
because of an "Unexpected problem".

Differential Revision: https://phabricator.services.mozilla.com/D185991

UltraBlame original commit: 9622edffb7e94deb655c7c7faf9238031868f9d8