Bug 1831324 - Part 1: Check if the module worker script has a parse error or an error to rethrow. r=jonco,dom-worker-reviewers,smaug
According to the spec,
https://html.spec.whatwg.org/#run-a-worker

14. In both cases, let onComplete given script be the following steps:
  1. If script is null or if script's error to rethrow is non-null, then:
    1. Queue a global task on the DOM manipulation task source given worker's relevant global object to fire an event named error at worker.
    2. Run the environment discarding steps for inside settings.
    3. Abort these steps.

Check if the module worker script has a parse error or an error to
rethrow. If it has, assign an error code to mRv and return false.

The error will be reported in CompileScriptRunnable
https://searchfox.org/mozilla-central/rev/892475f3ba2b959aeaef19d1d8602494e3f2ae32/dom/workers/WorkerPrivate.cpp#417

Differential Revision: https://phabricator.services.mozilla.com/D185531

UltraBlame original commit: 1695ada26bb4e92da7e7175a8fce65a01e86df5a