Bug 1824552 - Remove ScreenGetterWayland. r=stransky
This fixes screen sizes when using mutter with fractional scaling, and
shouldn't change behavior otherwise.

Without this patch, mutter with scale at 150% reports:

  [screen.width, screen.height] == [ 1920, 1080 ]

With this it reports the proper [ 2560, 1440 ].

The bug that caused us to introduce it was long fixed in gtk itself, as
I understand it, and ScreenGetterGtk is also what's used everywhere else
on Wayland.

Re-implement GetScreenForWindow in a way that works for ScreenGetterGtk
too, to avoid regressing any functionality on mutter.

Depends on D186341

Differential Revision: https://phabricator.services.mozilla.com/D186342

UltraBlame original commit: 3b5a0c666d81128f40fa48e3e30c16c64e64b959