Bug 1847636 [wpt PR 41370] - Fenced frame: Add WPT for loading an invalid FencedFrameConfig., a=testonly
Automatic update from web-platform-tests
Fenced frame: Add WPT for loading an invalid FencedFrameConfig.

Since FencedFrameConfigs can now be serialized and passed through
`window.postMessage()`, it is possible for a FencedFameConfig to be sent
to a context that can't load it. A fenced frame config is only valid
if it is in the same page that it was created in. (Each
FencedFrameURLMapping lives inside PageImpl).

This CL adds a test to make sure that if a FencedFrameConfig is sent
outside of the page it was created in, no navigation is allowed to
happen.

Change-Id: Iada09c6872a4206a024e63ebe75459738f394bbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4753440
Commit-Queue: Liam Brady <lbradygoogle.com>
Reviewed-by: Garrett Tanzer <gtanzerchromium.org>
Cr-Commit-Position: refs/heads/main{#1180845}

--

wpt-commits: 9fa3af6478f5f0dc4f04ee70588f5fb1527766f4
wpt-pr: 41370

UltraBlame original commit: c18ed1cfab096fb4cd9834cf467bb07d24b427eb