Bug 1840788: Adjust WPT percentage-heights-013.html to explicitly style its unrecognized-embed as a box that accepts CSS 'height'. r=TYLin
Browsers aren't currently interoperable on what the fallback rendering is for
embed elements with unrecognized 'type' attributes.  Firefox renders them as an
empty inline element (like an empty 'span'), whereas other browsers render them
as a fallback graphic with the 300x150 default replaced-element size.

We should probably try to achieve interop on this embed fallback behavior, but
it's not relevant to what this test is testing, which is whether percentage
heights can be resolved on children-of-flex-items. Percentages in fact *do*
resolve properly here, in all browsers.  But the last subtest is failing in
Firefox before this patch, simply because we're treating the embed as an empty
inline element -- and consequently, we're ignoring its 'height' property
entirely, just like we do on e.g. inline-level 'span'.

This patch avoids tripping over this 'embed' fallback-behavior behavioral
difference by forcing the embed element to be 'display: inline-block', which
makes its height property be recognized and honored, so that we can validate
that the percentage resovles.

Differential Revision: https://phabricator.services.mozilla.com/D186512

UltraBlame original commit: 8c2cc594742c1f00ced653bbf6d243c2686a41f9