Bug 1846191 clear mFakeAudioThread from the audio thread r=pehrsons
This makes the decision on whether to exit the audio thread loop consistent
with the decision to reset mFakeAudioThread.  Previously StopStream() would
reset mFakeAudioThread and wait for the audio thread to join, but the audio
thread loop would not exit if mFakeAudioThread was set to a new thread in a
parallel call to StartStream().

This patch does not limit audio thread lifetime to that of the MockCubeb.
There was no such guarantee previously: even a StopStream() call was not
sufficient because it did not block on concurrent StopStream() calls on the
same stream, which may have already cleared mFakeAudioThread but not yet
join()ed.

Depends on D185023

Differential Revision: https://phabricator.services.mozilla.com/D185024

UltraBlame original commit: 6c7a52e1f64e7dba4359758a1e181a6f37965a0a