Bug 1848908 - Force the module name used with wasm2c. r=firefox-build-system-reviewers,ahochheiden
On clang trunk, lld recently changed its behavior wrt linking wasm
files, whereas it now always puts a module name in the output, while
it didn't before. This leads wasm2c to use that module name in a way
that is not compatible with the rlbox code. Thus we force wasm2c to
use the module name it was defaulting to before.

Differential Revision: https://phabricator.services.mozilla.com/D186391

UltraBlame original commit: 44525f224544284d021a67ec424c9e3b46575009