Bug 1848744 [wpt PR 41462] - Adjust invalid subgrid behavior to match spec resolution, a=testonly
Automatic update from web-platform-tests
Adjust invalid subgrid behavior to match spec resolution

In https://chromium-review.googlesource.com/c/chromium/src/+/4647952,
I adjusted the computed style behavior for cases where `subgrid` is
specified on grid-template-columns/rows, but when the element is not
a valid subgrid. My interpretation at the time was that the computed
style should be `none` in this case.

I opened an issue to clarify
https://github.com/w3c/csswg-drafts/issues/9015, and I was given a
response after the above CL landed, which was to compute as if nothing
were specified, which means behaving like a grid (not subgrid), and
dropping the line names in the invalid subgrid definition. We were
mostly already doing this, but the computed style needed an extra
parameter to account for this, which this change adds. The test was
also updated to match the resolution, and now passes in both Chromium
and Firefox.

Bug: 1472916
Change-Id: I91b7de295adbdc1c5ee2dffe038fad390f58f4d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4779721
Reviewed-by: Ethan Jimenez <ethavarmicrosoft.com>
Reviewed-by: Alison Maher <almahermicrosoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmimicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1183761}

--

wpt-commits: 62fe5a0b7cdc176e8c599dcdd95c69c84dd3d79e
wpt-pr: 41462

UltraBlame original commit: e8763788444a357ffd49b580459f651955ff80cf