Bug 1850342 - Do not expose :-moz-broken to content. r=boris
This is not standard, and we don't use it internally (some chrome
stylesheets use it tho).

In the past this pseudo-class was more useful because it matched the
state for which <img> elements used an inline, but that's no longer
true, see bug 1196668 and co.

Depends on D186938

Differential Revision: https://phabricator.services.mozilla.com/D186939

UltraBlame original commit: 1b272e14e4029c03701c338b02b86323b4e02da9