Bug 1850238 - Set in-object-or-embed flag on image documents earlier. r=smaug
This fixes some intermittent that my previous patches introduce because
we don't reframe so often object or embed documents, and so the class
(which gets set on resize etc) isn't set.

The class is immutable in the object case so it's safe to just do this.

Differential Revision: https://phabricator.services.mozilla.com/D187132

UltraBlame original commit: d022a7b7801aafa0cfa1ced2b39409115f5544c4