Bug 1848991: Attach a RemoteAccessible to its parent before calling ProxyCreated. r=eeejay
This is what we did before the fix for bug 1779578.
Failure to do this meant that RemoteAccessible::IsLink() returned false even for a child of a HyperText Accessible, which meant the ATK object was created without the HyperLink interface.
This change requires that we move the handling of pending OOP child docs, since we can't add a child document until ProxyCreated has been called on the OuterDoc.

Differential Revision: https://phabricator.services.mozilla.com/D186423

UltraBlame original commit: 39fc3941397d2560d6d1b6ad33b987fc4c11edc9