Bug 1851166 - Avoid hard-coded white in osx global.css. r=desktop-theme-reviewers,sfoster
This shouldn't be drawn by default because of the appearance, but the
right color to use for -moz-DialogText is -moz-Dialog, not hard-coded
white.

I found this because this is what caused the macOS dark theme break so
awfully. The right fix for that is different, but this still looks worth
it.

Differential Revision: https://phabricator.services.mozilla.com/D187274

UltraBlame original commit: 52ca75cafa8057cd8a9bd121b74c64200af07640