Bug 1836805: test(webgpu): use explicit passing test case sections some CTS cases r=webgpu-reviewers,nical
Not every diff algorithm renders this as pure additions, but that's what
this change is! I recommend using the `patience` algorithm, if you have
it available to you. That way, it's easy to search for `^-` via regular
expressions, and prove it to yourself. 🙂

Differential Revision: https://phabricator.services.mozilla.com/D186577

UltraBlame original commit: 90fa7d45fa65b7fe15c71b795e26e1f177454b91