Bug 1802466 part 2: Change GetSelectionEndPointGeometry to handle vertical-writing-mode text nodes. r=emilio
This function had some code that we hadn't yet updated to support vertical
writing modes.  This patch makes it more generic so that it can properly handle
vertical-writing-mode text nodes (so that we can e.g. properly keep the caret
scrolled into view, in a vertical-writing-mode textfield).

I'm just focusing on vertical vs. horizontal generification for now.  I'm not
entirely sure if it makes sense to make it even-more-generic-still to support
reversed-inline-direction writing modes as well. I wasn't able to trigger any
issues with those in some brief testing, so I think this patch is sufficient
as-is, but I left a TODO(dholbert) code-comment to call out that ambiguity
that's still present in the code.

Differential Revision: https://phabricator.services.mozilla.com/D187160

UltraBlame original commit: 9461b0e136d35256a99a1e9ec21c31ae2f187590