Bug 1834690 - [devtools] Show custom properties value in computed panel. r=devtools-reviewers,ochameau.
At the moment, the (regular) computed properties PropertyView are created at
panel initialization, and then are shown/hidden depending of the selected element.
The list of properties is gathered from the panel's window computed property,
and not from the content page one.
That means that we can't build PropertyView for custom property at initialization
time, but potentially each time the computed property is refreshed.
In order to minimize dom mutations, when the view is refreshed, we:
- only remove the custom properties that are not set on the selected node
- update the value for the custom properties that are still set
- create new PropertyView for custom properties that weren't previously set

In order to keep the list ordered, we have to compute the index to put the new
elements in the right position. This is a bit tedious, but seems to work well
for performance.

A test is added for checking those custom properties.

Differential Revision: https://phabricator.services.mozilla.com/D182994

UltraBlame original commit: 2ac8ea15e2a7d1d6ca75d995a1d474f53af87c2e